Add custom build target
We need to use Rust nighly to build this

`cargo +nightly build --target x86_64-blog_os.json`

will now build our OS. It still doesn't do much!

We need to build `std` for our OS, so we need to use features that are
only available in nightly at the moment.

We set the panic strategy in the build target file, so we can remove it
from Cargo.toml.