Add global interface for Writer
We want to create a global writer that can be used as an interface from
other modules without carrying a `Writer` instance around.

We add the `lazy_static` and `spin` crates.

lazy_static allows us to declare a thing as only have a single
initialisation.

Spin allows us to gate access to that resource via a spinlock; see
https://en.wikipedia.org/wiki/Spinlock