Ensure QEMU exits from tests
We add the bootimage runner configuration for the test executables. This
allows us to pass a `-device` argument. QEMU supports a `isa-debug-exit`
device name which provides an easy way to exit QEMU from the guest
system. We use a generally unused [1] port of 0xf4 and 4 (0x04) bytes.

We add the x86_64 crate which abstracts the assembly required to use IO
ports.

We also configure the bootimage runner to map an exit code of 33 to 0.

[1] https://wiki.osdev.org/I/O_Ports#The_list