Make test output display in console
When we do `cargo test`, we want to see things in our shell; not printed
to QEMU.

A simple way to send data is via a serial port, so that's what we're
going to use here.

We add the uart_16550 crate to abstract using UARTs[1]. The common UARTs
today are all compatible to the 16550 UART [2], so we will use that
model for our testing framework.

Like the way we did VGA display, we wrap this serial port for reuse, and
add a couple of macros to simplify writing to it.

With this change, we now get output displaying test information in our
shell.

[1] https://en.wikipedia.org/wiki/Universal_asynchronous_receiver-transmitter
[2] https://en.wikipedia.org/wiki/16550_UART