Add integration test for printing after basic boot
This is very similar to an existing VGA buffer test. However, it's still
useful because it ensures that printing works with minimal setup. In
time, the OS and library might evolve to get more complex behaviour.
Ensuring printing works in the simplest environment possible is good for
our future selves when debugging!