Add test which should should panic
The test framework of the standard library supports a `#[should_panic]`
attribute that allows to construct tests that should fail. This is
useful for example to verify that a function fails when an invalid
argument is passed. Unfortunately this attribute isn't supported in
`#[no_std]` crates since it requires support from the standard library.

Here we offer a slightly less rich implementation. We add a panic
handler that exits with a success status code.

We then add a test case which has a failing assertion, and our
integration test is successful.