Simplify minimal integration tests
or integration tests that only have a single test function (like our
`should_panic` test), the test runner isn't really needed. For cases
like this, we can disable the test runner completely and run our test
directly in the `_start` function.

We do this by disabling the `harness` flag for that test. We can then
remove a lot of the test scaffolding code.