Add an interrupt module
We need to handle CPU Exceptions. These occur in various erroneous
situations, for example when accessing an invalid memory address or when
dividing by zero. To react to them we have to set up an interrupt
descriptor table that provides handler functions.

Here, we use the x86_64 crate again to provide data in the format
required by the hardware.

We also add a central place for initialization routines that can be
shared between the different `_start` functions in our `main.rs`,
`lib.rs`, and integration tests.

For interrupts, we use `lazy_static` again to define a global table. For
now we just handle breakpoints, and print when they happen. We call the
initialization from the new central place for such routines, and trigger
a breakpoint exception after it's configured.