Add deliberate unhandled CPU exception
We use unsafe to write to the invalid address `0xdeadbeef`. The virtual
address is not mapped to a physical address in the page tables, so a
page fault occurs. We haven't registered a page fault handler in our
Interrupt Descriptor Table (IDT), so a double fault occurs.

When we start our kernel now, we see that it enters an endless boot
loop. The reason for the boot loop is the following:

1. The CPU tries to write to `0xdeadbeef`, which causes a page fault.
2. The CPU looks at the corresponding entry in the IDT and sees that no
   handler function is specified. Thus, it can't call the page fault
   handler and a double fault occurs.
3. The CPU looks at the IDT entry of the double fault handler, but this
   entry does not specify a handler function either. Thus, a triple
   fault occurs.
4. A triple fault is fatal. QEMU reacts to it like most real hardware
   and issues a system reset.

So in order to prevent this triple fault, we need to either provide a
handler function for page faults or a double fault handler. We want to
avoid triple faults in all cases, so our next change will add a double
fault handler that is invoked for all unhandled exception types.