Add stack overflow handling
Our initial approach to double fault handling doesn't work for all
cases. Specifically, if our kernel overflows its stack and the guard
page is hit.

A guard page is a special memory page at the bottom of a stack that
makes it possible to detect stack overflows. The page is not mapped to
any physical frame, so accessing it causes a page fault instead of
silently corrupting other memory. The bootloader sets up a guard page
for our kernel stack, so a stack overflow causes a page fault.

When a page fault occurs the CPU looks up the page fault handler in the
IDT and tries to push the interrupt stack frame onto the stack. However,
the current stack pointer still points to the non-present guard page.
Thus, a second page fault occurs, which causes a double fault,
according to Table 8.3 in section 8.2.9 of
https://www.amd.com/system/files/TechDocs/24593.pdf#page=302

So the CPU tries to call the double fault handler now. However, on a
double fault the CPU tries to push the exception stack frame, too. The
stack pointer still points to the guard page, so a third page fault
occurs, which causes a triple fault and a system reboot. So our current
double fault handler can't avoid a triple fault in this case.

This change introduces using the hardware support for stack switching to
a known good stack when an exception occurs. The switching mechanism is
implemented as an Interrupt Stack Table (IST) which is a table of 7
pointers to known-good stacks. For each exception handlers, we can use a
stack in the IST. This switch happens before anything is pushed, so it
prevents the triple fault described above.

The IST is part of an old legacy structure called Task State Segment
(TSS). See https://en.wikipedia.org/wiki/Task_state_segment