Add hardware interrupt support
We introduce this using the 8259 Programmable Interrupt Controller (PIC)
interface. This is an older standard which is still supported for
backward-compatability reasons. There is of course an associated crate
which provides this functionality.

Hardware interrupts occur asynchronously, so they are completely
independent from the executed code and can occur at any time. Thus we
have a source of concurrency in our kernel now.

By default, PICs send interrupt vector numbers in the range 0 to 15 to
the CPU. These numbers are already occupied by CPU exceptions; for
example, 8 corresponds to a double fault. The first 32 interrupt vector
numbers are for CPU exceptions, so we remap the PIC interrupts to
different numbers. Here, we use 32 as the starting point.

We add a primary and a secondary PIC, as `ChainedPics`. This will be
eventually be configured something like:

```                     ____________                          ____________
Real Time Clock --> |            |   Timer -------------> |            |
ACPI -------------> |            |   Keyboard-----------> |            |      _____
Available --------> | Secondary  |----------------------> | Primary    |     |     |
Available --------> | Interrupt  |   Serial Port 2 -----> | Interrupt  |---> | CPU |
Mouse ------------> | Controller |   Serial Port 1 -----> | Controller |     |_____|
Co-Processor -----> |            |   Parallel Port 2/3 -> |            |
Primary ATA ------> |            |   Floppy disk -------> |            |
Secondary ATA ----> |____________|   Parallel Port 1----> |____________|
```

We can then initialise the PICs and enable interrupts. We add a timer
interrupt handler as our first example. The implementation is very
similar to the existing CPU exception handlers.