Send end of interrupt signal to the controller
This signal tells the controller that the interrupt was processed and
that the system is ready to receive the next interrupt. So the PIC
thinks we’re still busy processing the first timer interrupt and waits
patiently for the EOI signal before sending the next one.

With this change, a dot is printed each time the timer ticks.