Fix deadlock
If an interrupt handler tries to print whilst executed code is printing,
we can have a deadlock. The code from `_start` will have the lock for
the global `WRITER` and the interrupt handler will code will never
acquire the lock for the `WRITER` because it's already locked.

We can fix this by disabling intterupts as long as the Mutex is locked.