Fix broken test
We need to ensure that the current timer interrupt handler does not
interleave with our print test.

Again, we use `interrupts::without_interrupts` with a closure to make
that happen. We write a newline before the test string, since the
interrupt handler might have run before the test runs.

This fixes the data race that was causing this test to fail.