Use crate to handle keyboard
We had some fun handling it, but the mapping for scancodes to keys isn't
something that we want to spend much more time on.

Instead, we use the pc-keyboard crate and start off with a US keyboard
using scancode set 1. This will allow us to handle many more keys and
in my testing seems to be happy with '"', '@', and '#'.