Type-checking for page table access
Change the bootloader usage to allow the kernel access to page tables.

The bootloader sets up the page tables for our kernel. Here, we're
choosing to use the implementation which maps physical memory -
https://os.phil-opp.com/paging-implementation/#map-the-complete-physical-memory

We can then change our 'main' functions to use a macro which ensures we
don't have the wrong types on our functions.