Use OffsetPageTable to translate memory
Translating virtual to physical addresses is a common task in an OS
kernel, therefore the x86_64 crate provides an abstraction for it.

This change uses that abstraction and relies on our current
implementation that completely maps physical memory.

We can delete our `translate_addr` and `translate_addr_inner` functions
in favour of the things in the x86_64 crate.