Create a new mapping
We want to create a new mapping for a previously unmapped page. Here, we
map a given virtual page to `0xb8000`, the physical frame of the VGA
buffer.

This makes it easy for us to test if the mapping was created correctly;
we can write to the newly mapped page and see whether we see the write
appear on the screen.

We write `0x_f021_f077_f065_f04e` which represents the string "New!" on
white background.