Add Dummy allocator
We need to support dynamic memory allocation at runtime. We do this via
a heap which we can allocate and deallocate as needed. Rust manages
this using ownership and scope.

This implementation adds the interface, but always fails since it
returns a null pointer every time. Making this do something useful is
our next step.