Create a kernel heap
We want to create a heap memory region from which the allocator can
allocate memory. To do this, we need to define a virtual memory range
for the heap region and then map this region to physical frames. See
https://os.phil-opp.com/paging-introduction/ again for an overview of
virtual memory and page tables.

The first step is to define a virtual memory region for the heap. We can
choose any virtual address range that we like, as long as it is not
already used for a different memory region. Let's define it as the
memory starting at address `0x_4444_4444_0000` so that we can easily
recognize a heap pointer later. We set the heap size to 100 KiB for now.
If we need more space in the future, we can simply increase it

We create an `init_heap` function that maps the heap pages using the
`Mapper API` [1] that we introduced in the "Paging Implementation" [2]
post.

[1] https://os.phil-opp.com/paging-implementation/#using-offsetpagetable
[2] https://os.phil-opp.com/paging-implementation/