Use a simple bump allocator for memory
This linearly allocate memory by increasing ("bumping") a `next`
variable, which points to the start of the unused memory.

We use the same interface as the previous external
`linked_list_allocator`. This minimises the changes needed in calling
code.

Because of the requirements to have a static global allocator, we need
to use interior mutability [1] to lock it for mutual exclusion. This
safely turns a `&self` into a `&mut self` reference. We used a similar
approach for the VGA text buffer.

[1] https://doc.rust-lang.org/book/ch15-05-interior-mutability.html