Use async for keyboard interrupt handling
Keyboard handling is non-deterministic - it could happen at any time -
and latency-critical - the user should not perceive lag.

Here we change the keyboard interrupt handler to minimise the work done.
Rather than processing the keyboard input, it can read the scan code and
delegate the interpreting of that scan code to a background task.

We introduce `crossbeam_queue::ArrayQueue` as a lock-free buffer for
scan codes.

We use the `Stream` trait to define the task that yields multiple
asynchronous values.