Add executor with Waker support
The SimpleExecutor repeatedly polls for work, which isn't an efficient
use of the CPU.

In this change, we fix that. Our new Executor maintains a map of tasks
keyed by TaskId. We have a task queue of TaskIds and for now, we limit
that to 100 concurrent tasks. We also have a cache of Wakers for two
reasons: first, it improves performance by reusing the same waker for
multiple wake-ups of the same task instead of creating a new waker each
time. Second, it ensures that reference-counted wakers are not
deallocated inside interrupt handlers because it could lead to
deadlocks.

The CPU friendly approach comes with the new `sleep_if_idle` function in
the executors run-loop. We disable interrupts to ensure there are no
interrupts in between checking the task queue is empty and `hlt`-ing the
CPU. We atomically re-enable interrupts using the `enable_and_hlt`
function. This gives us much lower CPU utilisation.